requirements.txt
SpeechRecognition
gTTS
streamlit
openai
langdetect
sounddevice
numpy
pydub
bokeh
streamlit-mic-recorder
ffmpeg-python
streamlit-TTS
python-dotenv
cerebras_cloud_sdk